requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
proxy-tools==0.1.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywebview==6.0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
